@@ -20,7 +20,9 @@
         }
     ],
     "require": {
-        "php": ">=7.0.0"
+        "php": ">=8.2.0",
+        "ext-mbstring": "*",
+        "ext-zip": "*"
     },
     "autoload": {
         "files": [
